# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/backstage/menu": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Menu"
                ],
                "summary": "Menu View",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "int default",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            15,
                            20,
                            30,
                            40,
                            50
                        ],
                        "type": "integer",
                        "description": "int enums",
                        "name": "pageLimit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "id",
                            "key",
                            "url"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sortColumn",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "searchCategory",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.MenuViewListDTO"
                        }
                    }
                }
            }
        },
        "/backstage/menu/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Menu"
                ],
                "summary": "Menu Create",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.MenuCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/menu/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Menu"
                ],
                "summary": "Menu Delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/menu/edit/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Menu"
                ],
                "summary": "Menu Edit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.MenuCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/menu/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Menu"
                ],
                "summary": "Menu List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.MenuDTO"
                        }
                    }
                }
            }
        },
        "/backstage/menu/parent/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Menu"
                ],
                "summary": "Menu Parent List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.MenuDTO"
                        }
                    }
                }
            }
        },
        "/backstage/menu/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Menu"
                ],
                "summary": "Menu By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.MenuIdDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backstagedto.MenuCreateOrEditDTO": {
            "type": "object",
            "properties": {
                "feature": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "backstagedto.MenuDTO": {
            "type": "object",
            "properties": {
                "menu": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backstagedto.MenuNestData"
                    }
                }
            }
        },
        "backstagedto.MenuIdDTO": {
            "type": "object",
            "properties": {
                "menuById": {
                    "$ref": "#/definitions/backstagedto.MenuViewData"
                }
            }
        },
        "backstagedto.MenuNestData": {
            "type": "object",
            "properties": {
                "child": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backstagedto.MenuNestData"
                    }
                },
                "feature": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "backstagedto.MenuViewData": {
            "type": "object",
            "properties": {
                "feature": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "backstagedto.MenuViewListDTO": {
            "type": "object",
            "properties": {
                "menuViewList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backstagedto.MenuViewData"
                    }
                },
                "pageData": {
                    "$ref": "#/definitions/dto.PageForMultSearchDTO"
                }
            }
        },
...
        "dto.PageForMultSearchDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageLimit": {
                    "type": "integer"
                },
                "search": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "sort": {
                    "type": "string"
                },
                "sortColumn": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}